Fixes from online validator bugs (problematic copular forms)
Changes based on online validator that local validator missed:

First pass - most errors said: X is not an auxiliary verb in language

Step 1 taken to address this - restructured some copular construction
analyses to solve copula issues (swapped labelling of root and cop in
this instance)… rendering the label of each of these (Sin, Sen, Sé etc)
as cop.

But - after second run of validator - “sé' is not a copula in language “
Turns out the problem was lemmas need to be “Is” to be accepted as
copula in Irish.

Therefore, final changes are:
lemmas for copula forms that are  copula+det+pronoun = “is”
None of these copula forms are the root - instead they are ‘cop’ and
following NP is root.
All head values of dependents are changed accordingly (e.g punct to
root).

Other bugs addressed:
typo nsubj vs nusbj
Changed some problematic “Cop” forms to SCONJ (más) or substantive
nouns (cad).